factor out FieldLayoutCommunicator from FieldCommunicator2
